requirements.txt
numpy==1.19.2
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
PyOpenGL==3.1.1a1
